# Conflicts:
#	core/lib/config/src/configs/fri_prover.rs
#	core/lib/config/src/configs/house_keeper.rs
#	core/lib/config/src/testonly.rs
#	core/lib/env_config/src/fri_prover.rs
#	core/lib/env_config/src/house_keeper.rs
#	core/lib/protobuf_config/src/fri_prover.rs
#	core/lib/protobuf_config/src/house_keeper.rs
#	core/lib/protobuf_config/src/proto/fri_prover.proto
#	core/lib/protobuf_config/src/proto/house_keeper.proto
#	etc/env/base/house_keeper.toml

+++ core/lib/zksync_core/src/api_server/web3/namespaces/eth.rs
@@ -126,7 +126,7 @@ impl EthNamespace {
         let fee = self
             .state
             .tx_sender
-            .get_txs_fee_in_wei(tx.into(), scale_factor, acceptable_overestimation)
+            .get_txs_fee_in_wei(tx.into(), scale_factor, acceptable_overestimation as u64)
             .await?;
         Ok(fee.gas_limit)
     }
